extract storage account creation dialog as base class to support vm storage account createion as well.
